convert to modular Sinatra app
This attempts to fix an issue with the `request.body` not being a
StringIO instance, but wrapped depending on the application server
in use.

For Webrick, the raw request body seems not to be accessible anymore,
which is a shame. Switching to Puma helps a bit, but getting the raw
body still requires accessing an instance variabale from Rack::Lint...